Implementation of commands
This commit introduces the first “useful” version of this library.
It now includes a decent, but not perfect, representation of the
commands describen in control-spec.txt Version 1. Also it can generate
correct command string which can be sent to the control socket.

The introduced structures  will probably mostly be used for internal
purposes. Still missing is…

* … reply parsing and integration
* … useful “enduser” api for usage in applications